+++ jgiven-spring/src/main/java/com/tngtech/jgiven/integration/spring/JGivenSpringConfiguration.java
@@ -18,7 +18,7 @@ public class JGivenSpringConfiguration {
     }
 
     @Bean
-    public JGivenBeanFactoryPostProcessor jGivenPostBeanProcessor() {
+    public static JGivenBeanFactoryPostProcessor jGivenPostBeanProcessor() {
         return new JGivenBeanFactoryPostProcessor();
     }